Signed-off-by: CachorroC <juankpato87@gmail.com>
a Please enter the commit message for your changes. Lines starting
a with 'a' will be ignored, and an empty message aborts the commit.
a
a On branch weird-crow
a Your branch is up to date with 'origin/weird-crow'.
a
a Changes to be committed:
a	modified:   newCarpetasFinal.json
a	modified:   src/data/carpetas-raw.ts
a	modified:   src/models/demanda.ts
a